fix: Cast SUBSTRING position parameter to integer
PostgreSQL pg driver doesn't properly pass numeric parameters to
SUBSTRING FROM clause. Adding ::int cast fixes the issue.

+++ backend/src/services/rubro.service.ts
@@ -73,7 +73,7 @@ export class RubroService {
     if (prefijoChanged && existingPrefijo) {
       await execute(
         `UPDATE articulos
-         SET codigo = $1 || SUBSTRING(codigo FROM $2),
+         SET codigo = $1 || SUBSTRING(codigo FROM $2::int),
              updated_by = $3
          WHERE rubro_id = $4
            AND deleted_at IS NULL
